# backend/dist/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/businesses/{businessId}/services": {
      "post": {
        "operationId": "ServicesController_create",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Services"
        ]
      },
      "get": {
        "operationId": "ServicesController_findAll",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Services"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/services/{id}": {
      "get": {
        "operationId": "ServicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Services"
        ]
      },
      "patch": {
        "operationId": "ServicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Services"
        ]
      },
      "delete": {
        "operationId": "ServicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Services"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/staff": {
      "post": {
        "operationId": "StaffController_create",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Staff"
        ]
      },
      "get": {
        "operationId": "StaffController_findAll",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Staff"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/staff/{id}": {
      "get": {
        "operationId": "StaffController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Staff"
        ]
      },
      "patch": {
        "operationId": "StaffController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStaffDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Staff"
        ]
      },
      "delete": {
        "operationId": "StaffController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Staff"
        ]
      }
    },
...
    "/api/v1/businesses/{businessId}/notifications": {
      "post": {
        "operationId": "NotificationsController_create",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Notifications"
        ]
      },
      "get": {
        "operationId": "NotificationsController_findAll",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bookingId",
            "required": false,
            "in": "query",
            "description": "ID записи",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Тип уведомления",
            "schema": {
              "type": "string",
              "enum": [
                "SMS",
                "EMAIL",
                "VK"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Статус уведомления",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "SENT",
                "FAILED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/notifications/{id}": {
      "get": {
        "operationId": "NotificationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/notifications/{id}/resend": {
      "post": {
        "operationId": "NotificationsController_resend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/api/v1/businesses/{businessId}/notification-templates": {
      "get": {
        "operationId": "NotificationTemplatesController_getTemplates",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "NotificationTemplates"
        ]
      },
      "post": {
        "operationId": "NotificationTemplatesController_createTemplate",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "NotificationTemplates"
        ]
      }
    },
...
    "/api/v1/businesses/{businessId}/notification-templates/{templateId}": {
      "get": {
        "operationId": "NotificationTemplatesController_getTemplate",
        "parameters": [
          {
            "name": "templateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "NotificationTemplates"
        ]
      },
      "patch": {
        "operationId": "NotificationTemplatesController_updateTemplate",
        "parameters": [
          {
            "name": "templateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "NotificationTemplates"
        ]
      },
      "delete": {
        "operationId": "NotificationTemplatesController_deleteTemplate",
        "parameters": [
          {
            "name": "templateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "NotificationTemplates"
        ]
      }
    },
    "/api/v1/businesses/{businessId}/notification-templates/{templateId}/process": {
      "post": {
        "operationId": "NotificationTemplatesController_processTemplate",
        "parameters": [
          {
            "name": "templateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "NotificationTemplates"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mastro API",
    "description": "API для системы управления записями",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateServiceDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Название услуги"
          },
          "durationMinutes": {
            "type": "number",
            "description": "Длительность услуги в минутах",
            "minimum": 1
          },
          "price": {
            "type": "number",
            "description": "Цена услуги",
            "minimum": 0
          },
          "bufferBefore": {
            "type": "number",
            "description": "Буферное время до услуги в минутах",
            "minimum": 0
          },
          "bufferAfter": {
            "type": "number",
            "description": "Буферное время после услуги в минутах",
            "minimum": 0
          },
          "color": {
            "type": "string",
            "description": "Цвет услуги"
          }
        },
        "required": [
          "title",
          "durationMinutes",
          "price"
        ]
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Название услуги"
          },
          "durationMinutes": {
            "type": "number",
            "description": "Длительность услуги в минутах",
            "minimum": 1
          },
          "price": {
            "type": "number",
            "description": "Цена услуги",
            "minimum": 0
          },
          "bufferBefore": {
            "type": "number",
            "description": "Буферное время до услуги в минутах",
            "minimum": 0
          },
          "bufferAfter": {
            "type": "number",
            "description": "Буферное время после услуги в минутах",
            "minimum": 0
          },
          "color": {
            "type": "string",
            "description": "Цвет услуги"
          }
        }
      },
      "CreateStaffDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Имя сотрудника"
          },
          "phone": {
            "type": "string",
            "description": "Телефон сотрудника"
          },
          "serviceIds": {
            "description": "ID услуг, которые выполняет сотрудник",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateStaffDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Имя сотрудника"
          },
          "phone": {
            "type": "string",
            "description": "Телефон сотрудника"
          },
          "serviceIds": {
            "description": "ID услуг, которые выполняет сотрудник",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CreateNotificationDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "description": "ID записи"
          },
          "type": {
            "type": "string",
            "description": "Тип уведомления",
            "enum": [
              "SMS",
              "EMAIL",
              "VK"
            ]
          },
          "template": {
            "type": "string",
            "description": "Шаблон уведомления",
            "enum": [
              "BOOKING_CREATED",
              "BOOKING_CONFIRMED",
              "BOOKING_CANCELLED",
              "BOOKING_REMINDER"
            ]
          },
          "customMessage": {
            "type": "string",
            "description": "Пользовательское сообщение"
          }
        },
        "required": [
          "bookingId",
          "type",
          "template"
        ]
      },
...
      "CreateNotificationTemplateDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Тип шаблона",
            "enum": [
              "CONFIRM",
              "REMINDER",
              "CANCEL",
              "OTHER"
            ]
          },
          "channel": {
            "type": "string",
            "description": "Канал уведомления",
            "enum": [
              "SMS",
              "EMAIL",
              "VK"
            ]
          },
          "subject": {
            "type": "string",
            "description": "Тема сообщения"
          },
          "message": {
            "type": "string",
            "description": "Текст сообщения"
          },
          "isActive": {
            "type": "boolean",
            "description": "Активен ли шаблон"
          }
        },
        "required": [
          "type",
          "channel",
          "message"
        ]
      },
      "UpdateNotificationTemplateDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "description": "Тема сообщения"
          },
          "message": {
            "type": "string",
            "description": "Текст сообщения"
          },
          "isActive": {
            "type": "boolean",
            "description": "Активен ли шаблон"
          }
        }
      }
    }
  }
}